Year should always be the first 4 characters
The dateIssued_year index was empty when the dateIssued MODS field
only contains a year, like "2012". After some searching the function
substring-before(text(), '-') apparently returns an empty string if
the text() does not contain the needle ('-'). Change this to use the
substring(text(), 1, 4) function, because the contents of the
dateIssued MODS field should always start with the year as the first
4 characters, optionally followed by - and MONTH etc.